requirements.txt
appdirs==1.4.3
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.4
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.0
distro==1.4.0
dj-database-url==0.5.0
Django==3.1.1
django-cors-headers==3.5.0
django-environ==0.4.5
django-heroku==0.3.1
django-tinymce==3.0.2
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
django-filter==2.4.0
drf-yasg==1.17.1
gunicorn==20.0.4
html5lib==1.0.1
idna==2.8
inflection==0.5.1
ipaddr==2.2.0
isort==5.5.2
itypes==1.2.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
lockfile==0.12.2
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==0.6.2
psycopg2-binary
packaging==20.3
pep517==0.8.2
Pillow==7.2.0
pycodestyle==2.6.0
PyJWT==1.7.1
pylint==2.6.0
pyotp==2.4.0
pyparsing==2.4.6
pytoml==0.1.21
pytz==2020.1
requests==2.22.0
retrying==1.3.3
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.14.0
sqlparse==0.3.1
toml==0.10.1
uritemplate==3.0.1
urllib3==1.25.8
webencodings==0.5.1
whitenoise==5.2.0
wrapt==1.12.1
